+++ src/compare/compareArrayMissingItems.ts
@@ -1,13 +1,12 @@
-import { grayText } from "@vangware/forcli";
-import type { ReadOnlyObjectArray } from "@vangware/utils";
-import { INDENT } from "../constants";
 import { joinComma } from "../utils/joinComma";
+import { missingComment } from "../utils/missingComment";
 import { stringifyMap } from "../utils/stringifyMap";
 
 /**
  * Takes a source array and returns a string for missing items.
+ *
+ * @template Item Type of the items in the arrays being compared.
  * @param source Source array of missing items.
  */
-export const compareArrayMissingItems = <Item>(
-	source: ReadOnlyObjectArray<Item>
-) => grayText(`${INDENT}// Missing: ${joinComma(stringifyMap(source))}`);
+export const compareArrayMissingItems = <Item>(source: readonly Item[]) =>
+	missingComment(joinComma(stringifyMap(source)));