+++ src/suite/isSuiteResult.ts
@@ -1,18 +1,21 @@
 import { isArray, isObject, isString } from "@vangware/utils";
-import { isTestResult } from "../test/isTestResult";
+import { everyIsTestResult } from "../test/everyIsTestResult";
 import type { SuiteResult } from "../types/SuiteResult";
 
 /**
  * Check if given value is a `SuiteResult` object.
+ *
  * @param value Value to check.
  */
-export const isSuiteResult = (value: unknown): value is SuiteResult =>
+export const isSuiteResult = <Actual>(
+	value: Actual | SuiteResult
+): value is SuiteResult =>
 	isObject(value) &&
 	"name" in value &&
-	isString(value.name) &&
 	"failed" in value &&
-	isArray(value.failed) &&
-	value.failed.every(isTestResult) &&
 	"passed" in value &&
+	isString(value.name) &&
+	isArray(value.failed) &&
 	isArray(value.passed) &&
-	value.passed.every(isTestResult);
+	everyIsTestResult(value.failed) &&
+	everyIsTestResult(value.passed);