+++ src/suite/suite.ts
@@ -1,4 +1,3 @@
-import type { ReadOnlyObjectArray } from "@vangware/utils";
 import { testMap } from "../test/testMap";
 import { testResultsFilterTuple } from "../test/testResultsFilterTuple";
 import type { SuiteResult } from "../types/SuiteResult";
@@ -6,9 +5,11 @@ import type { Test } from "../types/Test";
 
 /**
  * Creates a new test suite (array of tests).
+ *
+ * @template Value The value being checked in the test.
  * @param tests List of test descriptions.
  */
-export const suite = <Value>(tests: ReadOnlyObjectArray<Test<Value>>) =>
+export const suite = <Value>(tests: readonly Test<Value>[]) =>
 	/**
 	 * @param name Name of the suite.
 	 */