+++ src/test/isTestResult.ts
@@ -3,11 +3,15 @@ import type { TestResult } from "../types/TestResult";
 
 /**
  * Check if given value is a `TestResult` object.
+ *
+ * @param Actual The actual type of the value.
  * @param value Value to check.
  */
-export const isTestResult = (value: unknown): value is TestResult =>
+export const isTestResult = <Actual>(
+	value: Actual | TestResult
+): value is TestResult =>
 	isObject(value) &&
 	"given" in value &&
-	isString(value.given) &&
 	"must" in value &&
+	isString(value.given) &&
 	isString(value.must);