+++ src/test/testResultToString.ts
@@ -4,6 +4,7 @@ import { testName } from "./testName";
 
 /**
  * Takes a `TestResult` and returns a stringified version of it.
+ *
  * @param testResult TestResult to generate the string from.
  */
 export const testResultToString = (testResult: TestResult) =>